refactor: move fetch logic into separate hook
