+++ src/components/CardList.tsx
@@ -1,6 +1,4 @@
-import axios, { AxiosError } from 'axios';
-import { useEffect, useState } from 'react';
-import { Character, Info } from '../Interfaces';
+import { useCharacters } from '../hooks/Characters';
 import Card from './Card';
 import ErrorMessage from './ErrorMessage';
 import Loader from './Loader';
@@ -10,36 +8,7 @@ interface IProps {
 }
 
 export default function CardList({ query }: IProps) {
-  const [characters, setCharacters] = useState<Character[]>([]);
-  const [loading, setLoading] = useState(false);
-  const [error, setError] = useState('');
-
-  useEffect(() => {
-    async function fetchProducts(name: string) {
-      setError('');
-      setLoading(true);
-      try {
-        const response = await axios.get<Info<Character[]>>(
-          'character' + (name ? `?name=${name}` : '')
-        );
-
-        setCharacters(response.data.results ?? []);
-      } catch (e: unknown) {
-        const error = e as AxiosError;
-
-        if (error.response?.status === 404) {
-          setCharacters([]);
-          return;
-        }
-
-        setError(error.message);
-      } finally {
-        setLoading(false);
-      }
-    }
-
-    fetchProducts(query);
-  }, [query]);
+  const { loading, error, characters } = useCharacters(query);
 
   return (
     <>
